Comment out most of what's left of external_startup_code.R from bdprobdiff
Half of this file was startup code that has been moved into other files.

The other half of this file is a list of source() calls to load nearly all
of the code for the project.  These are all being replaced with simply
including the files in the directory of the new package, since that will
automatically force them to be sourced when the package is built.
One issue with this is that all of the files have to be checked to make
sure that they do nothing other than defining functions.  For example,
some of them have a bit of testing code in them that was possible to run
at source time, though most of those bits were inside "if(FALSE)" protections.
Need to pull all those out and put them into proper tests in the package's
test directory.