Split tests and explorations out of gen_lognormal_overlay & move to tests area
The code split out into the tests area is a little different for this file.
Not everything in the tests looks to be a test.  Some of it looks like random
code for exploring a bit about the lognormal, e.g., building a big table of
lognormal values to possibly use as a way to get starting points for search
or as a simple lookup, etc.  May want to come back and pull this non-test
code out into more normal functions in the package.