Fix bad cross-reference link in documentation
The function linked to in the cross reference used to be part of the
R code for the package but is now in the code in the tests directory,
so the link no longer worked.  Not sure if there is a way to create a
link to a function in the tests directory, so I've just removed the
cross-link altogether.