Fix name conflict for "strength" in igraph and bipartite
Building bdpg kept giving the following message:
Warning message:
replacing previous import ‘bipartite::strength’ by ‘igraph::strength’ when loading ‘bdpg’

I had been importing the full bipartite and igraph libraries because there are so many calls to different routines in the packages.  It turns out that I was wrong about that for the bipartite package.  There ARE tons of calls for igraph functions, but only one call to a bipartite function.  So, I replaced the full import of the bipartite package with an import of only its function called networklevel.  I left the full import of the igraph package.  All seems to work fine now.