Comment out a noisy print statement during node linking & renamed bdpg-package.r to .R
The renaming doesn't show in git.  It doesn't seem to realize that the .r has gone to .R, so I'm just making a note of it here in a commit done just before I renamed it.  I renamed it because it wasn't being searched when I did grep commands over all the .R files in the directory and I do want it to be searched without having to make a more convoluted grep command to cover both upper and lower case.