Change gen_single_bdprob() to gen_single_bdprob_COR() and add extra elements to class
Expanding what's done in creating a problem so that it:
- adds a few new elements to the class structure
- computes the distribution and network metrics and saves them in the class
- saves the newly created problem to disk

These things were being done in the do_graph...() call before, if done at all.
Since all of these things contribute to the description of a problem,
independent of running a reserve selector, it makes more sense to do them when
the problem is created and store the information as part of the problem
description.  At some point, it might make more sense to pull some of these
externally computed attributes of the problem out into a separate class, but
not sure about that, so leaving it this way for now.