Add full_saved_bdprob_path to Xu_bd_problem class
Saving the full path to where the problem is written to disk just as a potential
clue for debugging and/or reproducibility.  It's not strictly necessary, but
there could be a point where it would be useful to have more clue about where
a problem found somewhere on disk came from.  Giving the full path could help
identify what run it came from, etc., if that sort of thing became lost but was
helpful to know.