Setting up and testing creation of reserve selection subdirectories
This is in transition and won't do a whole run properly.
It creates a nested list of directories but doesn't actually create the
directories since I'm just running some little tests on that right now.
Next step will be to reinstate the creation of the directories and to
move the act of creation of marxan directories to somewhere more
appropriate than in the original setup of a problem.