Change all references that are now related to Xu_prob_gen_info_class.
Searched for all related references and made appropriate changes where I knew what
to do.  In 2 cases, I punted for the moment.
- In add_error_to_spp...R, it creates the Apparent object, but that whole file is
going to have a bunch of changes since I'm changing how to deal with Apparent
problems.  So, I just marked a couple of places in that file where I will need to
pay attention to the prob gen info values when I start changing the Apparent stuff.
- In wrapping, it's not clear what values to put in the wrapper object.  For the
moment, I've just passed it the values from the problem it wraps.  However, I may
need to make a new class that holds both that information and the details about
how the wrapping was done, e.g., was it lognormal, if so, what parameters were used,
etc.  Will deal with that later, because I'm not sure yet how the information will
be used downstream as features in learning to predict.  When I know more about that,
then I can figure out what needs to be stored when the wrapping object is created.