Remove app_optimum_cost argument that looks superfluous
On looking closer at read_marxan_output_files(), it looks like app_optimum_cost
is computed inside the function rather than using a value that is passed in.
This makes more sense anyway since you don't know the app problem's view of
what is the app_optimum_cost until you've read the marxan output, which is
what's done in this routine.

I suspect that this argument came from using RStudio's "extract function" option
to gather up possible arguments in the function call and I made a mistake in not
removing it originally when I looked at the list of candidate arguments.