Move computation of spf_const into its own function
This species penalty factor is used by marxan as a control value and is
written to the network input, but it was embedded in writing the network to
the input files and the spf_const came back as a side effect.  It's clearer
to separate out the computation of the spf_const.