Replace directory creation with flat version that can be checksummed
Previous versions of these directory variables included the uuid of the problem
and that would be unique even if 2 problems were identical in every other way.
That means that 2 problems couldn't be compared for duplication by just
comparing their checksums.  Only storing the leaf name here will make it a little
more work to reference the directories because you'll have to build the whole
file path each time you need it, but it's worth it to be able to have a way to
easily compare whether 2 problems are identical up to uuid.