Make col and dir names constant values in object
I was going to make these be returns from a get_...() function since their
values are likely to always be the same in every problem object and therefore,
waste space through duplication.  However, I decided that since this is
research code that changes often, it might be better to waste the space in
favor of having the objects be self-documenting in terms of what they used
as headings and directories when they were built.  That way, an old object
whose structure had later changed could still be decoded if necessary.