Removed modification of prob object in gscp_13 on last commit so remove obj return
Previous commits removed the setting of derived_bdpg_dir_names slot on problem
since dirs are handled differently now and that slot no longer exists.
Modifying that slot was what forced having to return the problem object from
set_up_for_and_run_marxan...() before, so can now remove that from the function's
return.
That, in turn, means it's no longer to build and decode a return, so all of that
is removed in this commit.