Create dumping ground for some comments that may be useful for making docs
gscp_15 is filled with comments that are questions and code that has been
changed and commented out.  I want to delete these from the flow of the
code, but some of them have information that may be useful later, either
because I haven't properly fixed the code and it may be needed for debugging
my erroneous fix or because it describes some vectors and data frames
whose names aren't clear in their current context and these old bit of
commented code may help creating documentation for them later.