Add function to create global constants
I've added a ton of documentation at the start of this function that
aims to trick R into initializing something that seems like it should
be easy to do.
However, another reason to pay attention to this function is that the
documentation I've added uses all kinds of roxygen and .rd features,
so it's a good example to harvest things from if you can't figure out
how to do something.
I've also saved the roxygen code and output for all this documentation
in github bdpg issue #37 in case I delete it from here in some later
commit.