Add checksum to objects and provide function to compute it
Need a way to quickly compare two archived objects to see
whether they're effectively the same thing (i.e., up to
UUID and checksum fields).  This is because during
development, the same random seed and experiments are often
used over and over and we may need to quickly compare many
existing/archived complex objects so that we can
throw out duplicates.