Move wrapped prob search tracking file to dir of wrapped problem.
Tracking of details of the search for a wrapping distribution
were written to a file in the top directory of the tzar run
before the final directory for the wrapped problem was known.
That file should be stored with the wrapping instead, both for
ease of finding it and because it could be overwritten if more
than one wrapped problem is created in a tzar run.