Replace unnecessary use of solutions_df as first step to getting rid of it.
The solutions_df data frame is only used in this one place and that
use of it is unnecessary.  It looks like the solutions_df may be
vestigial and I may want to get rid of it, so this is a first step
toward that.  However, it does serve as a summary of the meanings of
various scattered vectors and that's useful.  Might want to just
comment it out but leave that comment around as a reminder of what
these values are.