Complete rewrite of gscp_15 to refactor into functions
gscp_15 still needs a lot of cleaning up since there is trash left all over
the file from the process of doing the major conversion.
The code at the very bottom of the file seems to work now and only some of
what occurs above it in the file should still be there.  The rest needs to
be deleted.