Separate building FP and FN matrices from applying them to create APP
Building the FP and FN matrices was done inside the apply_error...() function
even though it was only building those matrices one way.
Need the building of the matrices to be a function of whatever error model
is chosen, so that needs to be decoupled from applying the errors to the
occupancy data.

Note that this version runs to completion if you turn off the
"if (app_ct_error)  stop()" check in "apply_error...R", but fails at
that line if you leave it in.  Need to figure out what's going on there.