Fixes error where FP and FN rates always written out as 0.
There is a test in gscp_15...() that assigns FP and FN rates for
writing out based on whether the problem is a COR or an APP problem.
The test used to ask if the class of the problem was APP_prob_info_class,
but that is the class of the APP_prob_info attribute of the Xu_bd_problem
class.  It also tried to assign a value from APP_bd_prob@FP_const_rate
which is choosing the wrong thing again in the same way as above.
Now have a more direct test for an Apparen problem and assign FP, FN
values from one extra level of indirection to get the right value.