Break gscp_15 into smaller units.  Code still runs.
gscp_15 was huge and full of dead code, so the first step in
clarifying it is to split it into smaller piles of related code.
I have taken all the code that appears to be dead or no longer used
and put it into the gscp_15z...Rhold file since it looks like there
may still be important information in that code but it's currently
making it hard to see what is real code.  I've given it the Rhold
extension so that it's not executed during the Build as a way to
make sure that everything in there really is not needed to run the
program.  I rebuilt the library and ran a gen_4_variants test and
it all still seems to work ok, so I'm assuming that z code is no
longer needed.  There is one caution though:
do_marxan_analysis_and_output.R has some code that is hidden with
if(FALSE) and it calls some code that is in the 15z file, e.g.,
create_COR_master_output_structure().
I think that calling code is probably also dead at this point, but
I'm not ready to delete it until I know for sure and that will only
be after I've gone through this whole gscp_15 rewrite.