Add name of file to write to in calls to write_results_to_files()
This function always called the output files the same thing
(which was stored in a project.yaml variable,
parameters$summary_without_run_id_filename).  Because I am now
writing out chunks of the output in different csv files and then
recombining those later, it's more useful to have more informative
names on the files so that I know what's supposed to be in them.