Initialize APP_prob_info error rates to 0 instead of NA
I think that correct problems still have an APP_prob_info section
written out to results.  I think it just writes out the values that
are in the prototype.  The natural interpretation of an error
value for a correct would be 0, not NA, if trying to learn a
predictive model that included the error rate.  So, changing the
prototype values for all the errors to be 0 instead of NA.