Completely restructure save_rsprob_results_data() to handle Xu bench files.
The development of this function over time had not kept up with being
able to write out Xu benchmark files.  This file had old code for
doing that but it was way out of date and would not have worked if it
had been called.  In fact, the function to do it was never called
anywhere in bdpg.  Have now completely reorganized the
save_rsprob_results_data() function so that it handles both benchmark
and non-benchmark problems by refactoring the code into 3 new functions
that handle benchmark-only fields, non-benchmark-only fields, and
fields that both have in common.
I've left some of the old code in the file as dummy functions starting
with "hold___" in case they help explain some decision that I've made in
the refactoring.  I'll remove them on the next commit.