Remove old initializations in reading Xu benchmark file
Several values used to be set to NULL or NA before there was a
prototype for the problem object.  Now that there is a prototype,
R complains about type mismatch issues when the initial values are
set to NULL or NA during read_Xu_problem_from_Xu_file().

For example, R would stop and say something like:

  assignment of an object of class “logical” is not valid for
  @‘Xu_parameters’ in an object of class “PU_spp_pair_info_class”;
  is(value,

So, just removed the initializations in read_Xu_problem_from_Xu_file()
and now rely on the prototype to have initialized them instead.