Archiving marxan consistency check code
Removing this code for now because I don't think that it was
answering the question correctly and it's not important to bdpg.
It was only here to satisfy my curiosity about something that I
had read on the marxan mailing list.
Not sure I want to delete it completely though since it might be
useful at some point.  So, I'm putting it into a separate file
with some explanatory notes.