Change Xu_bd_problem@PU_spp_pair_indices to @cor_PU_spp_pair_indices
Both Cor and App problems make use of the PU_spp_pair_indices but
there was some confusion over which one was appropriate in certain
cases.
So, I've changed the name of the one in Xu_bd_problem class to
include "cor_" and the one in APP_prob_info_class to include
"app_".
This should also be done for num_spp and num_PUs since those have
a similar confusion in naming and use across cor and app.
Leaving them alone for the moment because they occur more frequently
and will take more work to change, but at the same time, almost
always have the same value so confusion is not such an issue.