Change message() to cat() and identify function called from
This line in calculate_mu() was spitting out a red output line on the
console that made me wonder if there was an error, plus the line was not
appearing in the output echoed to a file, so I'm changing it to
a cat() call instead.
Also, the output line didn't give any indication where it was being
output from, so I've added the name of the function in the call to
make it more informative.