Move random seed utilities out into new file and combine 2 funcs into 1
There are now 5 or 6 functions related to setting random seeds, so
I'm moving them out into their own file.
Part of the reason that there are so many functions now is that I have
combined 2 functions that always occurred together and made a new
function that calls them.  I have also taken the name for that function
from an existing function that I had thought before was the top-level
function in this seed stuff.  I changed the name of that function to
show its new helper status.