Moved apparently unused code to Hold area
When I ran Check, it was complaining about some undefined functions
that were only called in gscp_15b and the functions that called
them in gscp_15b were never called anywhere.
So, I think it's all vestigial code but I don't want to completely
chuck it in case I'm wrong.  So, I'm just moving it to the Hold
area for now.  Eventually, they can probably be chucked.