bdpg_gmail.json not allowed in R area by CHECK so move (and ignore)
This file was just used for some testing of running gmail from R code
but contained some sensitive authentication things, I think.
So, I moved it to the Hold area and put it in .gitignore so that it
wasn't saved anywhere outside my disk.
Also changed the reference to its path in the R code.