Fix problem with gen_wrapped_bdprob.R after pull's bad automerge
Couldn't push the most recent version of the code because it said:

    >>> git push origin refs/heads/master
    To https://github.com/langfob/bdpg.git
     ! [rejected]        master -> master (non-fast-forward)
    error: failed to push some refs to 'https://github.com/langfob/bdpg.git'
    hint: Updates were rejected because the tip of your current branch is     behind
    hint: its remote counterpart. Integrate the remote changes (e.g.
    hint: 'git pull ...') before pushing again.
    hint: See the 'Note about fast-forwards' in 'git push --help' for details.

So, I did a pull, which generated an automerge of gen_wrapped_bdproblem.R.
However, the result lost the most recent changes to gen_wrapped...R
So, I've copied the bad automerge result into a Hold directory in
case I've made a mistake and I've copied the good version of
gen_wrapped...R into the R directory to recommit.
Hopefully, that leaves everything as it was before the attempted push
but leaves me able to do a push now.