Change seed setting to also return and store R's .Random.seed array
When setting the random seed was not specifically called for in the
parameters list, the seed for each object was not recorded and the
only way to reproduce the object was to rerun the entire run that
generated it.
I've now added a new slot in the main objects to store R's internal
seed state array at each point where the seed is set or would be set
if the caller had asked for it.  That way, if you want to rebuild an
object but didn't explicitly set the seed, you can find what the
.Random.seed was when the object was created and restore its value
before recreating the object.
However, so far, there is no extra code for invoking that from the
parameters list.  It's more there as an emergency measure at the
moment.