Add ability to allow imperfect wraps
The optimization in choosing a lognormal to wrap around a Xu base problem
can generate distributions that don't properly wrap the base set of
species that occur on exactly 2 patches.  That is, the wrapping
distribution may have fewer species that occur on exactly 2 patches than
the Xu problem has.  It's not necessarily horrible if that happens;
it's just not a perfect wrap.  So, added a flag to determine whether
to allow that or not and to set the final distribution values correctly
if an imperfect wrap is allowed.