Clearing seed output for the moment
2017 12 19 - BTL
This is no longer correct since seeds can be set all over the place,
e.g., one for the whole tzar run, one for the rsrun, one for any
problems created, etc.  Need to rethink what should be written out to
the results files.  Probably need to have individual seed names
instead of one global seed output.
Also need a way to encode when the seed was not explicitly
set for the individual seeds.