Fix bug in setting of default value for bipartite_metrics_to_use
In the very recent changes to gscp_11c, I changed the way that some
of the flags were handled and ths allowed the bipartite_metrics_to_use
variable to end up being NULL when before it was defaulting to the
value I wanted, i.e., "ALLBUTDD".  I've fixed this now by testing
for NULL and having that cause the value to reset to the desired
default of "ALLBUTDD".