Remove all reference to unused global variables
bdpg created a global variable called GC that held global constant
values for things like column names with the original intention of
keeping them consistent but not having to pass them all over the place.
However, in the end, it was never used and raised an error every
time CHECK was run, so I'm removing all reference to the global stuff.