Add script for running tests of functions that are not exported
testthat::test_file() fails when you're testing functions that are not
exported from the package when you're running the tests outside of CHECK.
The stackoverflow question mentioned in this committed script file
suggests that you source the R file(s) containing the function(s) you
want to test.  That way, they end up in the namespace of the session
where you're working and they can be called.  I can't put this in the
tests area since it will end up getting run as another test if I do that,
so I'm putting it in the inst area until I know somewhere better for it.