Wrap as.character() around checksum to allow "==" tests to succeed
Tests for checksum equality of identical objects were failing in
test_gen_single_bdprob.R because of something weird about the object
that the checksum function (tools::md5sum) returns.
I expected the return to be a 32 character string holding the
checksum but it's some very strange object that contains a filename
followed by a newline, then the same number of blanks as the
length of the filename and then the 32 character checksum
surrounded by quotes.  At least that's the print representation
of what comes back.  If you wrap as.character() around that, it
turns into just the 32 characters of checksum and "==" works.
The problem with the original return value is that two objects
can have the same 32 character checksums but have been computed
on different temporary files and so the checksums fail to be "==".
I haven't been able to find any explanation of this anywhere,
so this is the best I can do to hack around it.