Add func to allow direct build of Xu prob w/o n, alpha, etc
Note: this version passes all CHECKS and tests.

Adding function so that you don't have to provide the 4 Xu
metaparameters that are used to derive the concrete parameters that
guide the building of the Xu problem.  This is useful because
while the Xu parameters are related to the theory behind finding
difficult problems, they aren't the values that are used in
building the problem itself, e.g., they don't directly tell you
how many links are possible.  That has to be derived.
When running experiments or doing some kind of search for
parameters that behave a certain way or doing experiments that
can't exceed a certain size, it may be easier to work with what
would be the derived parameters, directly.  For example, you may
want to set the number of nodes (PUs) or the number of links (spp)
directly.