Refactor out compute_target_num_links_between_2_groups_per_round()
This was inline code to compute the value of
target_num_links_between_2_groups_per_round.
This pulls that code out into a separate function that will be
further modified to correct some errors in the what it returns.
This version changes nothing about its behavior.  It just moves
the current code into a function and calls that function.