Decided to undo last change (about "|") since too much going on now
I think that the last change where I substituted double bars for
single bars is probably fine, but it was just a little side thing
that has nothing to do with what I'm working on now, so it could
end up leaving some deeply hidden bug if it's not right.  Need
to go back to it later.