Add variable setting temporarily needed for backwards compatibility
Had forgotten to add this to the wrapped test after adding it to
the single COR test.
It's currently needed because when the tests were originally written,
the default behavior in generating problems allowed there to be
a target number links between 2 groups per round of 0.  Now it's
a variable and defaults to requiring at least 1.  This means that
some of the old matches between generated objects don't match up
with values that I had figured out by hand when 0 was the target.
To get these tests to work with the new default, I have to go back
and do the hand figuring for results that have 1 as the target.