Fix bug introduced at commit c3c9fcb3 incorrectly assigning rsrun
Not sure why, but at that commit, I assigned the return value from
create_RSrun_dir_and_subdirs() to rsrun when before, it was not
assigned to anything.  This caused rsrun to be assigned a boolean
value rather than keeping its existing RSrun value.  This caused
a crash at the next line when save_rsrun() expected an RSrun value,
not a boolean.