Undoing previous commit - shouldn't have changed that name yet
Not sure if I'm ever going to want to include the reserve selector's
name in the results file name because it's already in the name of
the directory where the file is written.  If all results files
have the same name and just differ by directory, it may make code to
aggregate these files simpler, so I'm going back to the way I had it
before for now.