Hack test whether new & old versions of build_...() outputs match
This is a temporary hack to make sure that a simple input matrix
is transformed identically in the new and old versions of
build_PU_spp_pair_indices_from_occ_matrix() when switching from a
nested for() to using which().  This hack just adds a test routine
tb() and an extra argument to build_...() to control using old or
new version of the code.  The test works by just calling tb() but
I'm not testing the full bdpg with this version.  Now that I've
run tb() and seen that the old and new outputs are all.equal()
after they're sorted, I'm going to remove all this temporary stuff
and replace the for() with the which() in the real version of the
function.  I'm not going to keep tb() as a testthat test because
the for() version will no longer exist and so is no longer relevant.
This test is only a one-off to make sure that I've not messed up
the conversion from old to new form.