Fix bug in setting S_remaining_PUs_vec
This code shrinks the S_remaining_PUs_vec on each round, but was
addressing values in it using the PU_ID itself, which was the index
in the original vector, not the shrunken vector.
The fix finds the location of the PU_ID in the shrunken vector and
uses that index instead of the PU_ID to remove the element from the
vector (i.e., shrink it).