Refactor marxan etc at start of save_rsrun_results_data_for_one_rsrun()
Particularly note that I'm abstracting how the value is found for
app_solution_NUM_spp_covered__fromMarxan so that it's now computed
instead of derived from one of the marxan output values.
I'm doing this because it's unnecessary to do use the marxan output
value and I can make the code for doing this for marxan identical
to the way I do it for all reserve selectors.  In particular,
I'm not using the marxan value called:
marxan_output_values$marxan_mvbest_df$MPM

Below is the code that I've replaced with a call to a revised version
of compute_and_verify_APP_rep_scores_according_to_RS().  That function
also has different arguments now than it used to, since I'm now
doing the calculation of num_spp_covered inside that routine rather
than passing the value into it.

            #  NOTE:  Should probably replace this sum with a call to the
            #         more generic compute_num_spp_covered_by_solution()
            #         as a step toward abstracting this function more
            #         toward something that is the same for all reserve
            #         selectors.
            #  2018 01 26 - BTL
    app_solution_NUM_spp_covered__fromMarxan =
                                sum (marxan_output_values$marxan_mvbest_df$MPM)

    app_rep_scores_list_according_to_RS =
        compute_and_verify_APP_rep_scores_according_to_RS (
                                        app_solution_NUM_spp_covered__fromMarxan,
                                        COR_bd_prob@num_spp,
                                        "Marxan_SA")