Add flag to control whether to write network metrics to disk or not
Now that I'm adding a way to run graph metrics in batch, separately
from the generation of their associated problems, the directories
where metrics are written out during problem generation may not
exist during the batch run.  In that case, you only want to add the
values to the problem object and not write them to disk.