Fix inappropriate setting of spp_rep_targets to 1
This should be the default at the bottom of the call tree instead
of set along the way down like it was here.
As it was, it would have overwritten any other value that was
passed in and that would needlessly break any future run
where targets were not all 1.