Replace all COR_rs calls with APP_rs calls w/ 2 COR args
I was maintaining nearly identical functions for all kinds of
reserve selection calls but one had just the COR problem as the
argument and the other had the COR and the APP problems.
Have now replaced all of those with what was the APP version of each
except that the APP is removed from the name of each of the functions.
The old way was just a convenience thing to make sure that you always
called blah (COR, COR, ...) when doing COR problems.
Now I'm just being explicit so that I don't have to maintain 2 nearly
identical copies of all these functions.
of every one of these functions.