Fix bug in calc of euc_cost_rep_score that always returned NA
Using c() instead of list() in return from
build_and_write_rep_and_cm_scoress_list() flattens the result
into a single list rather than making it a list with 2 named
entries.  The downstream calculation of euc_COR_scores_list had
a bug when it mistakenly referenced the x$rep_scores_list element
of the return and that didn't exist because of the flattening.
That meant that the frac part of the euc calculation was always
NULL (because the referenced list element was not there) and
that meant the euc calculation always returned NA.