Fix gen_20... bug that caused combined errs 5 and 9 not to be generated
In cloning the cases originally, I forgot to make sure that both
FP and FN had non-zero values, so the cases that should have
generated BOTH FP & FN only had FP or FN set to a non-zero amount.
It seems like that should have made both matched and unmatched not
get generated, but I think that the old order of the if() statements
accidentally allowed the matched case to be generated.