Fix tests to handle wrap return values that changed to lists
When I changed wrapping to return values related to imperfect wrapping
back up the call change for writing out to results, that converted a
number of return values from non-lists to lists with named elements.
This broke some of the tests because they weren't pulling testable values
out of the lists, so now I have made the tests pull things out of the
named elements in the list.  None of the tests themselves were failing
other than in not extracting the test value from the list.  All tests
pass now.