Replace old call to get_marxan_best...() with its new version in dead spot
This call to get the marxan best solution is not currently used because of
the changes to do_marxan...(), but CHECK was complaining about it, and
rightly so.  For the moment, I'm just replacing it with the new, correct
call.  May want to remove this branch of the giant if() statement altogether,
but for the moment, it's not a problem.