Remove sqrt symbol from comment that was preventing latex from making pdfs
I had already added knitr chunk modifications earlier to prevent the
tidyverse library from generating messages that were crashing latex
because they contained the sqrt symbol.  However, when I added a comment
to mark those changes, I included the error message in the comment and
that error message contained the sqrt symbol.  Even though that symbol
was on a line that was commented out, latex was still trying to print
that comment in the output and choked on the sqrt symbol there.  So,
I had to replace the sqrt symbol in the comment with a bracketed phrase
indicating that's where the symbol used to be.  Now pdfs work fine.