Add parameters variable to load_existing_tzar...() call.
Not sure how this was working at all before, since it doesn't seem
like parameters should have been visible at all inside this function
or any that it called.  Some version of it might have just
accidentally been lying around in the global space from repeated
test runs.

I found that this was a problem when I did the first test of running
a simple greedy reserve selector on a loaded COR problem and it
couldn't build a file name correctly because it was missing the
parameters$fullOutputDir_NO_slash variable.