Remove unused code originally copied from single_action_using...()
The code in single_action_using_tzar_reps.R deals with graph metrics
in a different way (though that should probably change in the future)
and it allows generation of COR and APP and WRAP problems, which
I'm not doing in load_existing_tzar_run() now - though might allow
in the future.  So, getting rid of the cloned code that was related
to those actions.  Should have no effect here though, because all of
it was either commented out or wrapped in if(FALSE){} blocks.