Delete old version of saved net_bdprob before writing new one
In running graph metrics on problems from an old tzar run,
we need to replace the old version of the saved problem with a new
version that contains the network metric values just created.
So, to make sure that the old version really is replaced,
delete the old before writing the new one (since I'm not sure
whether the R save of an RDS file will overwrite or not).
Note that this does NOT replace the original version wherever it
was copied from in the original tzar run that has been reloaded.
It just does this inside the current run so that there is still
a record of the original run if needed.