Add err msg and stop_bdpg() when can't find allowable size Xu problem
This didn't fail before.  It just tried for the legal number of retries
and then returned whatever the last one tried was.  Now it will print a msg
saying it failed and how many times it tried, then it will quit.

Also adding a write statement to print the number of tries it took when it
does find a legal size problem in case I want to grep for these in the
log file later.