Fix bug in computing num_spp_to_generate for wrap
find_lognormal_to_wrap_around_Xu() was computing the inverse value of what
it was supposed to compute for the fraction to use in computing the
num_spp_te_generate in adding species to wrap around the base distribution.

In all previous runs, the input value in project.yaml for the
desired_Xu_spp_frac_of_all_spp had been set to 50%, so the bug had
no effect since the base and wrap proportions would both end up being 50%.
As soon as I changed the desired_Xu_spp_frac_of_all_spp to 25%, then
it was apparent that find_lognormal..() was doing a weird calculation that
just returned that desired_Xu_spp_frac_of_all_spp value as the value to
use for the wrap.  This meant that in the final wrap, 25% of the species
were wrap species and 75% were base species, exactly the opposite of the
intent.  Not sure where that old calculation came from, but it's fixed now.