Add ability to load random seed for run from yaml array
A small number of nectar runs are crashing because they run out of memory
when the bd problem is large and the nectar machine is one of the small
ones with only 4GB of memory.  In those cases, we want to be able to rerun
exactly the same problem on a larger machine.  To do this, we can just look
up the random seed that was set for the run that crashed and use that seed
again to start a run on a larger machine.  Since there are 8 or 10 of these
problems right now, it's more convenient to put them in a yaml array in
the project.yaml file and use that to choose the particular seed for a
particular run.  That's what this commit enables.