Clean up chunk options
I had thoght that "include=FALSE" was the chunk option to keep code from
being displayed, but it was "echo=FALSE" instead.  The "include" option
is for keeping chunk output out of the generated output, so I've removed
that from all places where it doesn't belong.  There ARE a couple of places
where you don't want the chunk output in the pdf, so those are left alone.